--- a/test/contrib/img_hash_test.dart
+++ b/test/contrib/img_hash_test.dart
@@ -44,7 +44,7 @@ void main() async {
     final img = cv.Mat.ones(256, 256, cv.MatType.CV_8UC1);
     for (var i = 0; i < img.rows; i++) {
       for (var j = 0; j < img.cols; j++) {
-        img.set(i, j, i + j);
+        img.set<cv.U8>(i, j, i + j);
       }
     }
     bmh.compute(img, hash);