--- a/test/core/mat_test.dart
+++ b/test/core/mat_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_print
+
 import 'dart:ffi' as ffi;
 
 import 'package:test/test.dart';
@@ -21,7 +23,7 @@ void main() async {
     final mat2 = cv.Mat.zeros(3, 3, cv.MatType.CV_8UC1);
     expect((mat2.width, mat2.height, mat2.channels), (3, 3, 1));
     expect(mat2.countNoneZero, equals(0));
-    mat2.set<int>(0, 0, 241);
+    mat2.setU8(0, 0, 241);
     expect(mat2.toList()[0][0], 241);
 
     final mat3 = cv.Mat.eye(3, 3, cv.MatType.CV_8UC3);
@@ -242,6 +244,55 @@ void main() async {
     expect((mat1.height, mat1.width, mat1.channels), (200, 100, 3));
   });
 
+  test('cv.Mat.data', () {
+    {
+      final mat = cv.Mat.zeros(3, 3, cv.MatType.CV_8UC3);
+      final data = mat.data;
+      expect(data.length, mat.rows * mat.cols * mat.channels);
+      expect(data[0], 0);
+      data[0] = 2;
+      data[1] = 4;
+      data[2] = 1;
+      expect(mat.at<cv.Vec3b>(0, 0), cv.Vec3b(2, 4, 1));
+    }
+
+    {
+      final src = cv.Mat.zeros(3840, 2160, cv.MatType.CV_16UC3);
+      final lut = List.generate(65536, (i) => 65535 - i);
+      final dst = src.clone();
+      final pSrc = src.data;
+      final pDst = dst.data;
+      final cn = src.channels;
+      final sw = Stopwatch();
+      sw.start();
+      for (int r = 0; r < src.rows; r++) {
+        for (int c = 0; c < src.cols; c++) {
+          for (int i = 0; i < cn; i++) {
+            final value = lut[pSrc[r + c * cn + i]];
+            pDst[r + c * cn + i] = value;
+          }
+        }
+      }
+      sw.stop();
+      // print('mat.data (${src.rows}x${src.cols}): ${sw.elapsedMilliseconds}ms');
+      expect(dst.at<cv.Vec3w>(0, 0), cv.Vec3w(65535, 65535, 65535));
+    }
+
+    {
+      final src = cv.Mat.zeros(3840, 2160, cv.MatType.CV_16UC3);
+      final dataList = List.generate(65536, (i) => 65535 - i);
+      final lut = cv.Mat.fromList(1, dataList.length, cv.MatType.CV_16UC1, dataList);
+      final sw = Stopwatch();
+      sw.start();
+      final dst = cv.LUT(src, lut);
+      sw.stop();
+      // print('mat.data LUT (${src.rows}x${src.cols}): ${sw.elapsedMilliseconds}ms');
+      expect(dst.isEmpty, false);
+      expect(dst.shape, src.shape);
+      expect(dst.at<cv.Vec3w>(0, 0), cv.Vec3w(65535, 65535, 65535));
+    }
+  });
+
   test('Mat test others', () {
     final mat0 = cv.Mat.fromScalar(200, 100, cv.MatType.CV_8UC3, cv.Scalar(1, 1, 1, 1));
     expect(mat0.props, equals([mat0.ptr.address]));
@@ -297,13 +348,16 @@ void main() async {
     expect(ptr0.address, greaterThan(0));
     expect(ptr0[0], 1);
     ptr0[0] = 21;
-    expect(mat.at<int>(0, 0), 21);
+    expect(mat.atI8(0, 0), 21);
     expect(ptr0[0], 21);
 
     final ptr1 = mat.ptrAt<cv.I8>(0, 0);
     expect(ptr1.address, greaterThan(0));
     expect(ptr1[0], 21);
     expect(List.generate(mat.cols, (i) => ptr1[i]), [21, 1, 1]);
+
+    mat.set<cv.I8>(0, 0, 2);
+    expect(mat.atI8(0, 0), 2);
   });
 
   test('Mat.ptrAt.U16', () {
@@ -390,7 +444,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec2b>(0, 0), cv.Vec2b(2, 4));
 
-    mat.set<int>(0, 0, 99);
+    mat.set<cv.U8>(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec2b>(0, 0, cv.Vec2b(99, 99));
@@ -400,7 +454,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec3b>(0, 0), cv.Vec3b(2, 4, 1));
 
-    mat.set<int>(0, 0, 99);
+    mat.setU8(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec3b>(0, 0, cv.Vec3b(99, 99, 99));
@@ -410,7 +464,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec4b>(0, 0), cv.Vec4b(2, 4, 1, 0));
 
-    mat.set<int>(0, 0, 99);
+    mat.setU8(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec4b>(0, 0, cv.Vec4b(99, 99, 99, 99));
@@ -422,7 +476,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec2w>(0, 0), cv.Vec2w(2, 4));
 
-    mat.set<int>(0, 0, 99);
+    mat.set<cv.U16>(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec2w>(0, 0, cv.Vec2w(99, 99));
@@ -432,7 +486,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec3w>(0, 0), cv.Vec3w(2, 4, 1));
 
-    mat.set<int>(0, 0, 99);
+    mat.setU16(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec3w>(0, 0, cv.Vec3w(99, 99, 99));
@@ -442,7 +496,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec4w>(0, 0), cv.Vec4w(2, 4, 1, 0));
 
-    mat.set<int>(0, 0, 99);
+    mat.setU16(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec4w>(0, 0, cv.Vec4w(99, 99, 99, 99));
@@ -454,7 +508,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec2s>(0, 0), cv.Vec2s(2, 4));
 
-    mat.set<int>(0, 0, 99);
+    mat.set<cv.I16>(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec2s>(0, 0, cv.Vec2s(99, 99));
@@ -464,7 +518,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec3s>(0, 0), cv.Vec3s(2, 4, 1));
 
-    mat.set<int>(0, 0, 99);
+    mat.setI16(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec3s>(0, 0, cv.Vec3s(99, 99, 99));
@@ -474,7 +528,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec4s>(0, 0), cv.Vec4s(2, 4, 1, 0));
 
-    mat.set<int>(0, 0, 99);
+    mat.setI16(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec4s>(0, 0, cv.Vec4s(99, 99, 99, 99));
@@ -486,7 +540,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec2i>(0, 0), cv.Vec2i(2, 4));
 
-    mat.set<int>(0, 0, 99);
+    mat.set<cv.I32>(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec2i>(0, 0, cv.Vec2i(99, 99));
@@ -496,7 +550,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec3i>(0, 0), cv.Vec3i(2, 4, 1));
 
-    mat.set<int>(0, 0, 99);
+    mat.setI32(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec3i>(0, 0, cv.Vec3i(99, 99, 99));
@@ -506,7 +560,7 @@ void main() async {
     expect(mat.at<int>(0, 0), 2);
     expect(mat.at<cv.Vec4i>(0, 0), cv.Vec4i(2, 4, 1, 0));
 
-    mat.set<int>(0, 0, 99);
+    mat.setI32(0, 0, 99);
     expect(mat.at<int>(0, 0), 99);
 
     mat.set<cv.Vec4i>(0, 0, cv.Vec4i(99, 99, 99, 99));
@@ -518,8 +572,8 @@ void main() async {
     expect(mat.at<double>(0, 0), closeTo(2, 1e-3));
     expect(mat.at<cv.Vec2f>(0, 0), cv.Vec2f(2, 4));
 
-    mat.set<double>(0, 0, 99);
-    expect(mat.at<double>(0, 0), 99);
+    mat.set<cv.F32>(0, 0, 99.0);
+    expect(mat.at<double>(0, 0), closeTo(99.0, 1e-6));
 
     mat.set<cv.Vec2f>(0, 0, cv.Vec2f(99, 99));
     expect(mat.at<cv.Vec2f>(0, 0), cv.Vec2f(99, 99));
@@ -528,8 +582,8 @@ void main() async {
     expect(mat.at<double>(0, 0), closeTo(2, 1e-3));
     expect(mat.at<cv.Vec3f>(0, 0), cv.Vec3f(2, 4, 1));
 
-    mat.set<double>(0, 0, 99);
-    expect(mat.at<double>(0, 0), 99);
+    mat.setF32(0, 0, 99.0);
+    expect(mat.at<double>(0, 0), closeTo(99.0, 1e-6));
 
     mat.set<cv.Vec3f>(0, 0, cv.Vec3f(99, 99, 99));
     expect(mat.at<cv.Vec3f>(0, 0), cv.Vec3f(99, 99, 99));
@@ -538,8 +592,8 @@ void main() async {
     expect(mat.at<double>(0, 0), closeTo(2, 1e-3));
     expect(mat.at<cv.Vec4f>(0, 0), cv.Vec4f(2, 4, 1, 0));
 
-    mat.set<double>(0, 0, 99);
-    expect(mat.at<double>(0, 0), 99);
+    mat.setF32(0, 0, 99.0);
+    expect(mat.at<double>(0, 0), closeTo(99.0, 1e-6));
 
     mat.set<cv.Vec4f>(0, 0, cv.Vec4f(99, 99, 99, 99));
     expect(mat.at<cv.Vec4f>(0, 0), cv.Vec4f(99, 99, 99, 99));
@@ -550,8 +604,8 @@ void main() async {
     expect(mat.at<double>(0, 0), closeTo(2, 1e-3));
     expect(mat.at<cv.Vec2d>(0, 0), cv.Vec2d(2, 4));
 
-    mat.set<double>(0, 0, 99);
-    expect(mat.at<double>(0, 0), 99);
+    mat.set<cv.F64>(0, 0, 99.0);
+    expect(mat.at<double>(0, 0), closeTo(99.0, 1e-6));
 
     mat.set<cv.Vec2d>(0, 0, cv.Vec2d(99, 99));
     expect(mat.at<cv.Vec2d>(0, 0), cv.Vec2d(99, 99));
@@ -560,8 +614,8 @@ void main() async {
     expect(mat.at<double>(0, 0), closeTo(2, 1e-3));
     expect(mat.at<cv.Vec3d>(0, 0), cv.Vec3d(2, 4, 1));
 
-    mat.set<double>(0, 0, 99);
-    expect(mat.at<double>(0, 0), 99);
+    mat.setF64(0, 0, 99.0);
+    expect(mat.at<double>(0, 0), closeTo(99.0, 1e-6));
 
     mat.set<cv.Vec3d>(0, 0, cv.Vec3d(99, 99, 99));
     expect(mat.at<cv.Vec3d>(0, 0), cv.Vec3d(99, 99, 99));
@@ -570,10 +624,63 @@ void main() async {
     expect(mat.at<double>(0, 0), closeTo(2, 1e-3));
     expect(mat.at<cv.Vec4d>(0, 0), cv.Vec4d(2, 4, 1, 0));
 
-    mat.set<double>(0, 0, 99);
-    expect(mat.at<double>(0, 0), 99);
+    mat.setF64(0, 0, 99.0);
+    expect(mat.at<double>(0, 0), closeTo(99.0, 1e-6));
 
     mat.set<cv.Vec4d>(0, 0, cv.Vec4d(99, 99, 99, 99));
     expect(mat.at<cv.Vec4d>(0, 0), cv.Vec4d(99, 99, 99, 99));
+  });
+
+  test('Mat at set perf', () {
+    final mat = cv.Mat.zeros(3840, 2160, cv.MatType.CV_8UC1);
+    final sw = Stopwatch();
+
+    {
+      sw.reset();
+      sw.start();
+      for (var row = 0; row < mat.rows; row++) {
+        for (var col = 0; col < mat.cols; col++) {
+          mat.at<int>(row, col);
+        }
+      }
+      sw.stop();
+      print('Mat(${mat.rows}, ${mat.cols}, ${mat.type}).at: ${sw.elapsedMilliseconds}ms');
+    }
+
+    {
+      sw.reset();
+      sw.start();
+      for (var row = 0; row < mat.rows; row++) {
+        for (var col = 0; col < mat.cols; col++) {
+          mat.atU8(row, col);
+        }
+      }
+      sw.stop();
+      print('Mat(${mat.rows}, ${mat.cols}, ${mat.type}).atU8: ${sw.elapsedMilliseconds}ms');
+    }
+
+    {
+      sw.reset();
+      sw.start();
+      for (var row = 0; row < mat.rows; row++) {
+        for (var col = 0; col < mat.cols; col++) {
+          mat.set<int>(row, col, 1);
+        }
+      }
+      sw.stop();
+      print('Mat(${mat.rows}, ${mat.cols}, ${mat.type}).set: ${sw.elapsedMilliseconds}ms');
+    }
+
+    {
+      sw.reset();
+      sw.start();
+      for (var row = 0; row < mat.rows; row++) {
+        for (var col = 0; col < mat.cols; col++) {
+          mat.setU8(row, col, 1);
+        }
+      }
+      sw.stop();
+      print('Mat(${mat.rows}, ${mat.cols}, ${mat.type}).setU8: ${sw.elapsedMilliseconds}ms');
+    }
   });
 }