--- a/test/video_test.dart
+++ b/test/video_test.dart
@@ -64,8 +64,8 @@ void main() async {
     expect(img.isEmpty, false);
     final testImg = cv.resize(img, (216, 216));
     final translationGround = cv.Mat.eye(2, 3, cv.MatType.CV_32FC1);
-    translationGround.set(0, 2, 11.4159);
-    translationGround.set(1, 2, 17.1828);
+    translationGround.set<double>(0, 2, 11.4159);
+    translationGround.setF64(1, 2, 17.1828);
 
     final wrappedImage = cv.warpAffine(
       testImg,