Releated:
- #53 
- #44 

simple perf test:
```bash
Mat(3840, 2160, CV_8UC1).at: 6937ms
Mat(3840, 2160, CV_8UC1).atU8: 4701ms
Mat(3840, 2160, CV_8UC1).set: 6188ms
Mat(3840, 2160, CV_8UC1).setU8: 3124ms
```

at: Mat.atU8, Mat.atI8, Mat.atU16, Mat.atI16, Mat.atI32, Mat.atF32, Mat.atF64

set: Mat.setU8, Mat.setI8, Mat.setU16, Mat.setI16, Mat.setI32, Mat.setF32, Mat.setF64

Mat.set support native types, e.g., Mat.set<cv.U8>()

In previous version, determine the Mat type is very expensive, but considering the mat type may change after creation, so it has to be get from native pointer, which costs lots of time, so for performance sensitive occasions, use atU8/setU8-like API.